Divide filter bar into 2 bars
